day3Coding done

+++ OnlineShoppingApp/src/main/java/com/masai/Service/AdminService.java
@@ -6,8 +6,8 @@ import com.masai.model.Product;
 
 public interface AdminService {
     public List<Product> getProductList();
-    public Product getProductById(Long id);
+    public Product getProductById(Integer id);
     public void addProduct(Product product);
     public void updateProduct(Product product);
-    public void removeProduct(Long id);
+    public void removeProduct(Integer id);
 }
